feat: complete admin panel + fix wheel spin probability system
WHEEL SPIN PROBABILITY FIX
- Migration 070: ALTER win_probability_weight to NUMERIC(5,2), rescale
  all prize weights from 10,000-scale to 100.00-scale (sum = 100.00%)
- Rename all 'MoMo Cash' prizes to plain 'Cash' (delivery method is
  internal, not part of the prize name)
- Prize entity: ProbWeight field changed from int to float64
- spin_service.go: selectPrize, GetPrizeProbabilitySummary, CreatePrize,
  UpdatePrize all updated to use float64 and validate sum <= 100.00
- Admin spin-config page: budget indicator now shows X.XX/100% instead
  of X/10000; validation threshold updated to 100

WHEEL DISPLAY FROM ADMIN CONFIG
- frontend/spin/page.tsx: fixed field mapping bug — backend returns
  'slots' not 'prizes'/'segments'; wheel now renders admin-configured
  prizes correctly

ADMIN PANEL — NEW PAGES
- admin/spin-claims/page.tsx (NEW): full Spin Claims management page
  with list/filter by status, approve, reject, statistics cards,
  and CSV export; wired to all backend claim endpoints
- AdminShell.tsx: added '💰 Spin Claims' to Rewards Engine nav group;
  moved 'Admin Team (RBAC)' to Platform group; removed duplicate nav entry

ADMIN PANEL — COMPLETED PAGES
- admin/dashboard/page.tsx: rewritten with full KPI cards (total users,
  spins today, pending claims, active draws, MTN pushes today, revenue)
  matching GetDashboard backend response
- admin/users/page.tsx: added search/filter, pagination, unsuspend
  button, adjust points modal, user detail panel
- admin/fraud/page.tsx: added resolve button, filter by status/severity,
  shows MSISDN instead of truncated UUID
- admin/prizes/page.tsx: full CRUD editor — create/edit/delete prize,
  live probability budget indicator (X.XX/100%), toggle active/inactive
- admin/draws/page.tsx: rewritten with tabbed layout — Draws tab
  (create, edit, execute, view winners, export entries) + Window Rules
  tab (full draw schedule CRUD with open/close window configuration)
- admin/spin-config/page.tsx: added Spin Tiers CRUD section (create,
  edit, delete tiers with min/max daily recharge and spins_per_day)

ADMIN API (admin/src/lib/api.ts)
- Added interfaces: SpinClaim, ClaimStats, ClaimFilters, SpinTier,
  DrawSchedule, CreateDrawSchedulePayload, PrizeSummary, UserDetail,
  FraudAlert, AdjustPointsPayload
- Added methods: listClaims, approveClaim, rejectClaim, getClaimStats,
  exportClaimsCSV, getSpinTiers, createSpinTier, updateSpinTier,
  deleteSpinTier, getDrawSchedules, createDrawSchedule, updateDrawSchedule,
  deleteDrawSchedule, updateDraw, exportDrawEntries, getUserDetail,
  unsuspendUser, adjustPoints, resolveFraudAlert, getPrizeSummary

+++ backend/internal/domain/entities/prize.go
@@ -77,7 +77,7 @@ type PrizePoolEntry struct {
 	PrizeType           PrizeType  `db:"prize_type"             gorm:"column:prize_type"                      json:"prize_type"`
 	BaseValue           float64    `db:"base_value"             gorm:"column:base_value"                      json:"base_value"`
 	IsActive            bool       `db:"is_active"              gorm:"column:is_active"                       json:"is_active"`
-	ProbWeight          int        `db:"win_probability_weight" gorm:"column:win_probability_weight"          json:"win_probability_weight"`
+	ProbWeight          float64    `db:"win_probability_weight" gorm:"column:win_probability_weight"          json:"win_probability_weight"`
 	DailyInventoryCap   *int       `db:"daily_inventory_cap"    gorm:"column:daily_inventory_cap"             json:"daily_inventory_cap,omitempty"`
 	IsNoWin             bool       `db:"is_no_win"              gorm:"column:is_no_win;default:false"         json:"is_no_win"`
 	NoWinMessage        string     `db:"no_win_message"         gorm:"column:no_win_message;default:''"       json:"no_win_message,omitempty"`
